+++ src/main/java/com/nagornov/CorporateMessenger/application/controller/UserProfilePhotoController.java
@@ -3,15 +3,19 @@ package com.nagornov.CorporateMessenger.application.controller;
 import com.nagornov.CorporateMessenger.application.dto.common.FileRequest;
 import com.nagornov.CorporateMessenger.application.dto.common.HttpResponse;
 import com.nagornov.CorporateMessenger.application.applicationService.UserPhotoApplicationService;
-import com.nagornov.CorporateMessenger.domain.annotation.ant.ValidUuid;
+import com.nagornov.CorporateMessenger.domain.exception.BindingErrorException;
 import com.nagornov.CorporateMessenger.domain.model.user.UserPhoto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.UUID;
+
+@Validated
 @RestController
 @RequiredArgsConstructor
 public class UserProfilePhotoController {
@@ -23,9 +27,11 @@ public class UserProfilePhotoController {
             consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-    ResponseEntity<UserPhoto> loadUserPhoto(@Validated @ModelAttribute FileRequest request) {
-        UserPhoto response =
-                userPhotoApplicationService.loadUserPhoto(request);
+    ResponseEntity<UserPhoto> loadUserPhoto(@ModelAttribute FileRequest request, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new BindingErrorException("FileRequest validation error", bindingResult);
+        }
+        UserPhoto response = userPhotoApplicationService.loadUserPhoto(request);
         return ResponseEntity.status(201).body(response);
     }
 
@@ -34,9 +40,11 @@ public class UserProfilePhotoController {
             value = "/api/user/{userId}/photo/main",
             produces = MediaType.IMAGE_JPEG_VALUE
     )
-    ResponseEntity<Resource> getMainUserPhotoByUserId(@ValidUuid @PathVariable("userId") String userId) {
-        Resource response =
-                userPhotoApplicationService.getMainUserPhotoByUserId(userId);
+    ResponseEntity<Resource> getMainUserPhotoByUserId(@PathVariable UUID userId, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new BindingErrorException("PathVariable(userId) validation error", bindingResult);
+        }
+        Resource response = userPhotoApplicationService.getMainUserPhotoByUserId(userId);
         return ResponseEntity.status(200).body(response);
     }
 
@@ -45,9 +53,11 @@ public class UserProfilePhotoController {
             value = "/api/user/photo/{photoId}",
             produces = MediaType.IMAGE_JPEG_VALUE
     )
-    ResponseEntity<Resource> getUserPhotoByPhotoId(@ValidUuid @PathVariable("photoId") String photoId) {
-        Resource response =
-                userPhotoApplicationService.getUserPhotoByPhotoId(photoId);
+    ResponseEntity<Resource> getUserPhotoByPhotoId(@PathVariable UUID photoId, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new BindingErrorException("PathVariable(photoId) validation error", bindingResult);
+        }
+        Resource response = userPhotoApplicationService.getUserPhotoByPhotoId(photoId);
         return ResponseEntity.status(200).body(response);
     }
 
@@ -56,9 +66,11 @@ public class UserProfilePhotoController {
             value = "/api/user/photo/{photoId}",
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-    ResponseEntity<UserPhoto> setMainUserPhoto(@ValidUuid @PathVariable("photoId") String photoId) {
-        UserPhoto response =
-                userPhotoApplicationService.setMainUserPhoto(photoId);
+    ResponseEntity<UserPhoto> setMainUserPhoto(@PathVariable UUID photoId, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new BindingErrorException("PathVariable(photoId) validation error", bindingResult);
+        }
+        UserPhoto response = userPhotoApplicationService.setMainUserPhoto(photoId);
         return ResponseEntity.status(200).body(response);
     }
 
@@ -67,9 +79,11 @@ public class UserProfilePhotoController {
             value = "/api/user/photo/{photoId}",
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-    ResponseEntity<HttpResponse> deleteUserPhoto(@ValidUuid @PathVariable("photoId") String photoId) {
-        HttpResponse response =
-                userPhotoApplicationService.deleteUserPhoto(photoId);
+    ResponseEntity<HttpResponse> deleteUserPhoto(@PathVariable UUID photoId, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new BindingErrorException("PathVariable(photoId) validation error", bindingResult);
+        }
+        HttpResponse response = userPhotoApplicationService.deleteUserPhoto(photoId);
         return ResponseEntity.status(200).body(response);
     }
 }
